resume-roaster/requirements.txt
pdfplumber>=0.9.0
python-docx>=1.1.0
anthropic>=0.34.0
rich>=13.0.0
python-dotenv>=1.0.0
